apps/api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
pydantic-settings==2.8.1
python-jose[cryptography]==3.4.0
requests==2.32.3
httpx==0.28.1
yfinance==0.2.57
feedparser==6.0.11
google-genai==1.13.0
faiss-cpu==1.10.0
numpy==2.2.5
python-multipart==0.0.20
